Moved all consts to a single file
* Exported consts (ABIs and contract addresses) from MainnetHelper
* Also fixed a filename typo (MainnerHelper->MainnetHelper)
