+++ test/MainnetHelper.js
@@ -1,0 +1,139 @@
+const {BigNumber, FixedFormat, FixedNumber, formatFixed, parseFixed} = require("@ethersproject/bignumber");
+const { expect } = require("chai");
+
+// grab the private api key from the private repo
+require('dotenv').config({ path: 'secrets/alchemy.env' })
+let alchemy_url = "https://eth-mainnet.alchemyapi.io/v2/" + process.env.ALCHEMY_API_KEY;
+
+/* CONTRACT ADDRESSES ON MAINNET */
+const CURVE_TRIPOOL2_ADDRESS = "0xd51a44d3fae010294c616388b506acda1bfaae46";
+const USDT_ADDRESS = "0xdAC17F958D2ee523a2206206994597C13D831ec7";
+const WETH9_ADDRESS = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2";
+const CURVE_FACTORY_ADDRESS = "0xB9fC157394Af804a3578134A6585C0dc9cc990d4";
+const CURVE_3POOL_ADDRESS = "0xbEbc44782C7dB0a1A60Cb6fe97d0b483032FF1C7";
+const OUSD_SWAP_ADDRESS = "0xcecad69d7d4ed6d52efcfa028af8732f27e08f70";
+const OUSD_ADDRESS = "0x2A8e1E676Ec238d8A992307B495b45B3fEAa5e86";
+
+const USDT_POOL_INDEX_NUMBER = 0
+const WETH_POOL_INDEX_NUMBER = 2
+
+/* ABIs */
+const OUSD_ABI = [{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"owner","type":"address"},{"indexed":true,"internalType":"address","name":"spender","type":"address"},{"indexed":false,"internalType":"uint256","name":"value","type":"uint256"}],"name":"Approval","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"previousGovernor","type":"address"},{"indexed":true,"internalType":"address","name":"newGovernor","type":"address"}],"name":"GovernorshipTransferred","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"previousGovernor","type":"address"},{"indexed":true,"internalType":"address","name":"newGovernor","type":"address"}],"name":"PendingGovernorshipTransfer","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"totalSupply","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"rebasingCredits","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"rebasingCreditsPerToken","type":"uint256"}],"name":"TotalSupplyUpdatedHighres","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"from","type":"address"},{"indexed":true,"internalType":"address","name":"to","type":"address"},{"indexed":false,"internalType":"uint256","name":"value","type":"uint256"}],"name":"Transfer","type":"event"},{"inputs":[],"name":"_totalSupply","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_owner","type":"address"},{"internalType":"address","name":"_spender","type":"address"}],"name":"allowance","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_spender","type":"address"},{"internalType":"uint256","name":"_value","type":"uint256"}],"name":"approve","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"_account","type":"address"}],"name":"balanceOf","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"account","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"burn","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"uint256","name":"_newTotalSupply","type":"uint256"}],"name":"changeSupply","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"claimGovernance","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"_account","type":"address"}],"name":"creditsBalanceOf","outputs":[{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_account","type":"address"}],"name":"creditsBalanceOfHighres","outputs":[{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"decimals","outputs":[{"internalType":"uint8","name":"","type":"uint8"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_spender","type":"address"},{"internalType":"uint256","name":"_subtractedValue","type":"uint256"}],"name":"decreaseAllowance","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"governor","outputs":[{"internalType":"address","name":"","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_spender","type":"address"},{"internalType":"uint256","name":"_addedValue","type":"uint256"}],"name":"increaseAllowance","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"string","name":"_nameArg","type":"string"},{"internalType":"string","name":"_symbolArg","type":"string"},{"internalType":"address","name":"_vaultAddress","type":"address"}],"name":"initialize","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"isGovernor","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"isUpgraded","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_account","type":"address"},{"internalType":"uint256","name":"_amount","type":"uint256"}],"name":"mint","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"name","outputs":[{"internalType":"string","name":"","type":"string"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"nonRebasingCreditsPerToken","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"nonRebasingSupply","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"rebaseOptIn","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"rebaseOptOut","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"rebaseState","outputs":[{"internalType":"enum OUSD.RebaseOptions","name":"","type":"uint8"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"rebasingCredits","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"rebasingCreditsHighres","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"rebasingCreditsPerToken","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"rebasingCreditsPerTokenHighres","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"symbol","outputs":[{"internalType":"string","name":"","type":"string"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"totalSupply","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_to","type":"address"},{"internalType":"uint256","name":"_value","type":"uint256"}],"name":"transfer","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"_from","type":"address"},{"internalType":"address","name":"_to","type":"address"},{"internalType":"uint256","name":"_value","type":"uint256"}],"name":"transferFrom","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"_newGovernor","type":"address"}],"name":"transferGovernance","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"vaultAddress","outputs":[{"internalType":"address","name":"","type":"address"}],"stateMutability":"view","type":"function"}]
+const OUSD_SWAP_ABI = [{"constant":true,"inputs":[],"name":"governor","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"buyOusdWithUsdt","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"buyOusdWithDai","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[],"name":"claimGovernance","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[],"name":"withdrawAll","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"sellOusdForDai","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"buyOusdWithUsdc","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"sellOusdForUsdc","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"isGovernor","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"sellOusdForUsdt","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"_newGovernor","type":"address"}],"name":"transferGovernance","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"token","type":"address"},{"internalType":"uint256","name":"amount","type":"uint256"}],"name":"withdraw","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[],"name":"rebaseOptIn","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"inputs":[],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"previousGovernor","type":"address"},{"indexed":true,"internalType":"address","name":"newGovernor","type":"address"}],"name":"PendingGovernorshipTransfer","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"internalType":"address","name":"previousGovernor","type":"address"},{"indexed":true,"internalType":"address","name":"newGovernor","type":"address"}],"name":"GovernorshipTransferred","type":"event"}]
+const CURVE_TRIPOOL2_ABI = [{"anonymous":false,"inputs":[{"indexed":true,"name":"buyer","type":"address"},{"indexed":false,"name":"sold_id","type":"uint256"},{"indexed":false,"name":"tokens_sold","type":"uint256"},{"indexed":false,"name":"bought_id","type":"uint256"},{"indexed":false,"name":"tokens_bought","type":"uint256"}],"name":"TokenExchange","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"provider","type":"address"},{"indexed":false,"name":"token_amounts","type":"uint256[3]"},{"indexed":false,"name":"fee","type":"uint256"},{"indexed":false,"name":"token_supply","type":"uint256"}],"name":"AddLiquidity","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"provider","type":"address"},{"indexed":false,"name":"token_amounts","type":"uint256[3]"},{"indexed":false,"name":"token_supply","type":"uint256"}],"name":"RemoveLiquidity","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"provider","type":"address"},{"indexed":false,"name":"token_amount","type":"uint256"},{"indexed":false,"name":"coin_index","type":"uint256"},{"indexed":false,"name":"coin_amount","type":"uint256"}],"name":"RemoveLiquidityOne","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"deadline","type":"uint256"},{"indexed":true,"name":"admin","type":"address"}],"name":"CommitNewAdmin","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"admin","type":"address"}],"name":"NewAdmin","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"deadline","type":"uint256"},{"indexed":false,"name":"admin_fee","type":"uint256"},{"indexed":false,"name":"mid_fee","type":"uint256"},{"indexed":false,"name":"out_fee","type":"uint256"},{"indexed":false,"name":"fee_gamma","type":"uint256"},{"indexed":false,"name":"allowed_extra_profit","type":"uint256"},{"indexed":false,"name":"adjustment_step","type":"uint256"},{"indexed":false,"name":"ma_half_time","type":"uint256"}],"name":"CommitNewParameters","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"admin_fee","type":"uint256"},{"indexed":false,"name":"mid_fee","type":"uint256"},{"indexed":false,"name":"out_fee","type":"uint256"},{"indexed":false,"name":"fee_gamma","type":"uint256"},{"indexed":false,"name":"allowed_extra_profit","type":"uint256"},{"indexed":false,"name":"adjustment_step","type":"uint256"},{"indexed":false,"name":"ma_half_time","type":"uint256"}],"name":"NewParameters","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"initial_A","type":"uint256"},{"indexed":false,"name":"future_A","type":"uint256"},{"indexed":false,"name":"initial_gamma","type":"uint256"},{"indexed":false,"name":"future_gamma","type":"uint256"},{"indexed":false,"name":"initial_time","type":"uint256"},{"indexed":false,"name":"future_time","type":"uint256"}],"name":"RampAgamma","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"current_A","type":"uint256"},{"indexed":false,"name":"current_gamma","type":"uint256"},{"indexed":false,"name":"time","type":"uint256"}],"name":"StopRampA","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"admin","type":"address"},{"indexed":false,"name":"tokens","type":"uint256"}],"name":"ClaimAdminFee","type":"event"},{"inputs":[{"name":"owner","type":"address"},{"name":"admin_fee_receiver","type":"address"},{"name":"A","type":"uint256"},{"name":"gamma","type":"uint256"},{"name":"mid_fee","type":"uint256"},{"name":"out_fee","type":"uint256"},{"name":"allowed_extra_profit","type":"uint256"},{"name":"fee_gamma","type":"uint256"},{"name":"adjustment_step","type":"uint256"},{"name":"admin_fee","type":"uint256"},{"name":"ma_half_time","type":"uint256"},{"name":"initial_prices","type":"uint256[2]"}],"outputs":[],"stateMutability":"nonpayable","type":"constructor"},{"stateMutability":"payable","type":"fallback"},{"gas":3361,"inputs":[{"name":"k","type":"uint256"}],"name":"price_oracle","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3391,"inputs":[{"name":"k","type":"uint256"}],"name":"price_scale","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3421,"inputs":[{"name":"k","type":"uint256"}],"name":"last_prices","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":468,"inputs":[],"name":"token","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":582,"inputs":[{"name":"i","type":"uint256"}],"name":"coins","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":597,"inputs":[],"name":"A","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":11991,"inputs":[],"name":"gamma","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":21673,"inputs":[],"name":"fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":11096,"inputs":[{"name":"xp","type":"uint256[3]"}],"name":"fee_calc","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":11582,"inputs":[],"name":"get_virtual_price","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"inputs":[{"name":"i","type":"uint256"},{"name":"j","type":"uint256"},{"name":"dx","type":"uint256"},{"name":"min_dy","type":"uint256"}],"name":"exchange","outputs":[],"stateMutability":"payable","type":"function"},{"inputs":[{"name":"i","type":"uint256"},{"name":"j","type":"uint256"},{"name":"dx","type":"uint256"},{"name":"min_dy","type":"uint256"},{"name":"use_eth","type":"bool"}],"name":"exchange1","outputs":[],"stateMutability":"payable","type":"function"},{"gas":3122,"inputs":[{"name":"i","type":"uint256"},{"name":"j","type":"uint256"},{"name":"dx","type":"uint256"}],"name":"get_dy","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":26582,"inputs":[{"name":"amounts","type":"uint256[3]"},{"name":"xp","type":"uint256[3]"}],"name":"calc_token_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":738687,"inputs":[{"name":"amounts","type":"uint256[3]"},{"name":"min_mint_amount","type":"uint256"}],"name":"add_liquidity","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":233981,"inputs":[{"name":"_amount","type":"uint256"},{"name":"min_amounts","type":"uint256[3]"}],"name":"remove_liquidity","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":3429,"inputs":[{"name":"amounts","type":"uint256[3]"},{"name":"deposit","type":"bool"}],"name":"calc_token_amount","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":13432,"inputs":[{"name":"token_amount","type":"uint256"},{"name":"i","type":"uint256"}],"name":"calc_withdraw_one_coin","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":648579,"inputs":[{"name":"token_amount","type":"uint256"},{"name":"i","type":"uint256"},{"name":"min_amount","type":"uint256"}],"name":"remove_liquidity_one_coin","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":389808,"inputs":[],"name":"claim_admin_fees","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":163102,"inputs":[{"name":"future_A","type":"uint256"},{"name":"future_gamma","type":"uint256"},{"name":"future_time","type":"uint256"}],"name":"ramp_A_gamma","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":157247,"inputs":[],"name":"stop_ramp_A_gamma","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":306190,"inputs":[{"name":"_new_mid_fee","type":"uint256"},{"name":"_new_out_fee","type":"uint256"},{"name":"_new_admin_fee","type":"uint256"},{"name":"_new_fee_gamma","type":"uint256"},{"name":"_new_allowed_extra_profit","type":"uint256"},{"name":"_new_adjustment_step","type":"uint256"},{"name":"_new_ma_half_time","type":"uint256"}],"name":"commit_new_parameters","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":683438,"inputs":[],"name":"apply_new_parameters","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":23222,"inputs":[],"name":"revert_new_parameters","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":77260,"inputs":[{"name":"_owner","type":"address"}],"name":"commit_transfer_ownership","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":65937,"inputs":[],"name":"apply_transfer_ownership","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":23312,"inputs":[],"name":"revert_transfer_ownership","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":40535,"inputs":[],"name":"kill_me","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":23372,"inputs":[],"name":"unkill_me","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":38505,"inputs":[{"name":"_admin_fee_receiver","type":"address"}],"name":"set_admin_fee_receiver","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":3378,"inputs":[],"name":"last_prices_timestamp","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3408,"inputs":[],"name":"initial_A_gamma","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3438,"inputs":[],"name":"future_A_gamma","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3468,"inputs":[],"name":"initial_A_gamma_time","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3498,"inputs":[],"name":"future_A_gamma_time","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3528,"inputs":[],"name":"allowed_extra_profit","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3558,"inputs":[],"name":"future_allowed_extra_profit","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3588,"inputs":[],"name":"fee_gamma","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3618,"inputs":[],"name":"future_fee_gamma","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3648,"inputs":[],"name":"adjustment_step","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3678,"inputs":[],"name":"future_adjustment_step","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3708,"inputs":[],"name":"ma_half_time","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3738,"inputs":[],"name":"future_ma_half_time","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3768,"inputs":[],"name":"mid_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3798,"inputs":[],"name":"out_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3828,"inputs":[],"name":"admin_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3858,"inputs":[],"name":"future_mid_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3888,"inputs":[],"name":"future_out_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3918,"inputs":[],"name":"future_admin_fee","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4057,"inputs":[{"name":"arg0","type":"uint256"}],"name":"balances","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3978,"inputs":[],"name":"D","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4008,"inputs":[],"name":"owner","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":4038,"inputs":[],"name":"future_owner","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":4068,"inputs":[],"name":"xcp_profit","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4098,"inputs":[],"name":"xcp_profit_a","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4128,"inputs":[],"name":"virtual_price","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4158,"inputs":[],"name":"is_killed","outputs":[{"name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"gas":4188,"inputs":[],"name":"kill_deadline","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4218,"inputs":[],"name":"transfer_ownership_deadline","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4248,"inputs":[],"name":"admin_actions_deadline","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":4278,"inputs":[],"name":"admin_fee_receiver","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"}]
+const USDT_ABI = [{"constant":true,"inputs":[],"name":"name","outputs":[{"name":"","type":"string"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"_upgradedAddress","type":"address"}],"name":"deprecate","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"_spender","type":"address"},{"name":"_value","type":"uint256"}],"name":"approve","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"deprecated","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"_evilUser","type":"address"}],"name":"addBlackList","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"totalSupply","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"_from","type":"address"},{"name":"_to","type":"address"},{"name":"_value","type":"uint256"}],"name":"transferFrom","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"upgradedAddress","outputs":[{"name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"}],"name":"balances","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"decimals","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"maximumFee","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"_totalSupply","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[],"name":"unpause","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"name":"_maker","type":"address"}],"name":"getBlackListStatus","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"},{"name":"","type":"address"}],"name":"allowed","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"paused","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"name":"who","type":"address"}],"name":"balanceOf","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[],"name":"pause","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"getOwner","outputs":[{"name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"owner","outputs":[{"name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"symbol","outputs":[{"name":"","type":"string"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"_to","type":"address"},{"name":"_value","type":"uint256"}],"name":"transfer","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"newBasisPoints","type":"uint256"},{"name":"newMaxFee","type":"uint256"}],"name":"setParams","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"amount","type":"uint256"}],"name":"issue","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"amount","type":"uint256"}],"name":"redeem","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"name":"_owner","type":"address"},{"name":"_spender","type":"address"}],"name":"allowance","outputs":[{"name":"remaining","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"basisPointsRate","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"}],"name":"isBlackListed","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"_clearedUser","type":"address"}],"name":"removeBlackList","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"MAX_UINT","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"newOwner","type":"address"}],"name":"transferOwnership","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"_blackListedUser","type":"address"}],"name":"destroyBlackFunds","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"inputs":[{"name":"_initialSupply","type":"uint256"},{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_decimals","type":"uint256"}],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"name":"amount","type":"uint256"}],"name":"Issue","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"amount","type":"uint256"}],"name":"Redeem","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"newAddress","type":"address"}],"name":"Deprecate","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"feeBasisPoints","type":"uint256"},{"indexed":false,"name":"maxFee","type":"uint256"}],"name":"Params","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"_blackListedUser","type":"address"},{"indexed":false,"name":"_balance","type":"uint256"}],"name":"DestroyedBlackFunds","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"_user","type":"address"}],"name":"AddedBlackList","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"_user","type":"address"}],"name":"RemovedBlackList","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"owner","type":"address"},{"indexed":true,"name":"spender","type":"address"},{"indexed":false,"name":"value","type":"uint256"}],"name":"Approval","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"from","type":"address"},{"indexed":true,"name":"to","type":"address"},{"indexed":false,"name":"value","type":"uint256"}],"name":"Transfer","type":"event"},{"anonymous":false,"inputs":[],"name":"Pause","type":"event"},{"anonymous":false,"inputs":[],"name":"Unpause","type":"event"}]
+const WETH9_ABI = [{"constant":true,"inputs":[],"name":"name","outputs":[{"name":"","type":"string"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"guy","type":"address"},{"name":"wad","type":"uint256"}],"name":"approve","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"totalSupply","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"src","type":"address"},{"name":"dst","type":"address"},{"name":"wad","type":"uint256"}],"name":"transferFrom","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"name":"wad","type":"uint256"}],"name":"withdraw","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[],"name":"decimals","outputs":[{"name":"","type":"uint8"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"}],"name":"balanceOf","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"symbol","outputs":[{"name":"","type":"string"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"name":"dst","type":"address"},{"name":"wad","type":"uint256"}],"name":"transfer","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[],"name":"deposit","outputs":[],"payable":true,"stateMutability":"payable","type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"},{"name":"","type":"address"}],"name":"allowance","outputs":[{"name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"payable":true,"stateMutability":"payable","type":"fallback"},{"anonymous":false,"inputs":[{"indexed":true,"name":"src","type":"address"},{"indexed":true,"name":"guy","type":"address"},{"indexed":false,"name":"wad","type":"uint256"}],"name":"Approval","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"src","type":"address"},{"indexed":true,"name":"dst","type":"address"},{"indexed":false,"name":"wad","type":"uint256"}],"name":"Transfer","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"dst","type":"address"},{"indexed":false,"name":"wad","type":"uint256"}],"name":"Deposit","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"src","type":"address"},{"indexed":false,"name":"wad","type":"uint256"}],"name":"Withdrawal","type":"event"}]
+const CURVE_FACTORY_ABI = [{"anonymous":false,"inputs":[{"indexed":false,"name":"base_pool","type":"address"}],"name":"BasePoolAdded","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"coins","type":"address[4]"},{"indexed":false,"name":"A","type":"uint256"},{"indexed":false,"name":"fee","type":"uint256"},{"indexed":false,"name":"deployer","type":"address"}],"name":"PlainPoolDeployed","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"coin","type":"address"},{"indexed":false,"name":"base_pool","type":"address"},{"indexed":false,"name":"A","type":"uint256"},{"indexed":false,"name":"fee","type":"uint256"},{"indexed":false,"name":"deployer","type":"address"}],"name":"MetaPoolDeployed","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"pool","type":"address"},{"indexed":false,"name":"gauge","type":"address"}],"name":"LiquidityGaugeDeployed","type":"event"},{"inputs":[{"name":"_fee_receiver","type":"address"}],"outputs":[],"stateMutability":"nonpayable","type":"constructor"},{"gas":21716,"inputs":[{"name":"_base_pool","type":"address"}],"name":"metapool_implementations","outputs":[{"name":"","type":"address[10]"}],"stateMutability":"view","type":"function"},{"inputs":[{"name":"_from","type":"address"},{"name":"_to","type":"address"}],"name":"find_pool_for_coins","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[{"name":"_from","type":"address"},{"name":"_to","type":"address"},{"name":"i","type":"uint256"}],"name":"find_pool_for_coins1","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":2663,"inputs":[{"name":"_pool","type":"address"}],"name":"get_base_pool","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":2699,"inputs":[{"name":"_pool","type":"address"}],"name":"get_n_coins","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":5201,"inputs":[{"name":"_pool","type":"address"}],"name":"get_meta_n_coins","outputs":[{"name":"","type":"uint256"},{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":9164,"inputs":[{"name":"_pool","type":"address"}],"name":"get_coins","outputs":[{"name":"","type":"address[4]"}],"stateMutability":"view","type":"function"},{"gas":21345,"inputs":[{"name":"_pool","type":"address"}],"name":"get_underlying_coins","outputs":[{"name":"","type":"address[8]"}],"stateMutability":"view","type":"function"},{"gas":20185,"inputs":[{"name":"_pool","type":"address"}],"name":"get_decimals","outputs":[{"name":"","type":"uint256[4]"}],"stateMutability":"view","type":"function"},{"gas":19730,"inputs":[{"name":"_pool","type":"address"}],"name":"get_underlying_decimals","outputs":[{"name":"","type":"uint256[8]"}],"stateMutability":"view","type":"function"},{"gas":5281,"inputs":[{"name":"_pool","type":"address"}],"name":"get_metapool_rates","outputs":[{"name":"","type":"uint256[2]"}],"stateMutability":"view","type":"function"},{"gas":20435,"inputs":[{"name":"_pool","type":"address"}],"name":"get_balances","outputs":[{"name":"","type":"uint256[4]"}],"stateMutability":"view","type":"function"},{"gas":39733,"inputs":[{"name":"_pool","type":"address"}],"name":"get_underlying_balances","outputs":[{"name":"","type":"uint256[8]"}],"stateMutability":"view","type":"function"},{"gas":3135,"inputs":[{"name":"_pool","type":"address"}],"name":"get_A","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":5821,"inputs":[{"name":"_pool","type":"address"}],"name":"get_fees","outputs":[{"name":"","type":"uint256"},{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":13535,"inputs":[{"name":"_pool","type":"address"}],"name":"get_admin_balances","outputs":[{"name":"","type":"uint256[4]"}],"stateMutability":"view","type":"function"},{"gas":33407,"inputs":[{"name":"_pool","type":"address"},{"name":"_from","type":"address"},{"name":"_to","type":"address"}],"name":"get_coin_indices","outputs":[{"name":"","type":"int128"},{"name":"","type":"int128"},{"name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"gas":3089,"inputs":[{"name":"_pool","type":"address"}],"name":"get_gauge","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3119,"inputs":[{"name":"_pool","type":"address"}],"name":"get_implementation_address","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3152,"inputs":[{"name":"_pool","type":"address"}],"name":"is_meta","outputs":[{"name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"gas":5450,"inputs":[{"name":"_pool","type":"address"}],"name":"get_pool_asset_type","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":5480,"inputs":[{"name":"_pool","type":"address"}],"name":"get_fee_receiver","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_coins","type":"address[4]"},{"name":"_A","type":"uint256"},{"name":"_fee","type":"uint256"}],"name":"deploy_plain_pool","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_coins","type":"address[4]"},{"name":"_A","type":"uint256"},{"name":"_fee","type":"uint256"},{"name":"_asset_type","type":"uint256"}],"name":"deploy_plain_pool","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_coins","type":"address[4]"},{"name":"_A","type":"uint256"},{"name":"_fee","type":"uint256"},{"name":"_asset_type","type":"uint256"},{"name":"_implementation_idx","type":"uint256"}],"name":"deploy_plain_pool","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"name":"_base_pool","type":"address"},{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_coin","type":"address"},{"name":"_A","type":"uint256"},{"name":"_fee","type":"uint256"}],"name":"deploy_metapool","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"name":"_base_pool","type":"address"},{"name":"_name","type":"string"},{"name":"_symbol","type":"string"},{"name":"_coin","type":"address"},{"name":"_A","type":"uint256"},{"name":"_fee","type":"uint256"},{"name":"_implementation_idx","type":"uint256"}],"name":"deploy_metapool1","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"gas":93079,"inputs":[{"name":"_pool","type":"address"}],"name":"deploy_gauge","outputs":[{"name":"","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"gas":1206132,"inputs":[{"name":"_base_pool","type":"address"},{"name":"_fee_receiver","type":"address"},{"name":"_asset_type","type":"uint256"},{"name":"_implementations","type":"address[10]"}],"name":"add_base_pool","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":382072,"inputs":[{"name":"_base_pool","type":"address"},{"name":"_implementations","type":"address[10]"}],"name":"set_metapool_implementations","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":379687,"inputs":[{"name":"_n_coins","type":"uint256"},{"name":"_implementations","type":"address[10]"}],"name":"set_plain_implementations","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":38355,"inputs":[{"name":"_gauge_implementation","type":"address"}],"name":"set_gauge_implementation","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":1139545,"inputs":[{"name":"_pools","type":"address[32]"},{"name":"_asset_types","type":"uint256[32]"}],"name":"batch_set_pool_asset_type","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":38415,"inputs":[{"name":"_addr","type":"address"}],"name":"commit_transfer_ownership","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":58366,"inputs":[],"name":"accept_transfer_ownership","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":40996,"inputs":[{"name":"_manager","type":"address"}],"name":"set_manager","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":38770,"inputs":[{"name":"_base_pool","type":"address"},{"name":"_fee_receiver","type":"address"}],"name":"set_fee_receiver","outputs":[],"stateMutability":"nonpayable","type":"function"},{"gas":12880,"inputs":[],"name":"convert_metapool_fees","outputs":[{"name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"gas":8610792,"inputs":[{"name":"_pools","type":"address[10]"}],"name":"add_existing_metapools","outputs":[{"name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"},{"gas":3438,"inputs":[],"name":"admin","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3468,"inputs":[],"name":"future_admin","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3498,"inputs":[],"name":"manager","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3573,"inputs":[{"name":"arg0","type":"uint256"}],"name":"pool_list","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3558,"inputs":[],"name":"pool_count","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3633,"inputs":[{"name":"arg0","type":"uint256"}],"name":"base_pool_list","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3618,"inputs":[],"name":"base_pool_count","outputs":[{"name":"","type":"uint256"}],"stateMutability":"view","type":"function"},{"gas":3863,"inputs":[{"name":"arg0","type":"address"}],"name":"base_pool_assets","outputs":[{"name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"gas":3838,"inputs":[{"name":"arg0","type":"uint256"},{"name":"arg1","type":"uint256"}],"name":"plain_implementations","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"},{"gas":3708,"inputs":[],"name":"gauge_implementation","outputs":[{"name":"","type":"address"}],"stateMutability":"view","type":"function"}]
+
+module.exports = {
+
+    /* helper functions */
+    helperResetNetwork: async function (lockBlock) {
+
+        // Reset hardhat mainnet fork	
+	    await network.provider.request({
+	    	method: "hardhat_reset",
+		    params: [
+    			{
+				    forking: {
+					    jsonRpcUrl: alchemy_url,
+					    blockNumber: lockBlock,
+    				},
+			    },
+		    ],
+	    });
+    },
+
+    /*
+        Fork is starting us with plenty of ETH so
+        1. Convert ETH to WETH (because this is what Curve is working with)
+        2. WETH->USDT on TriCrypto2@Curve
+        3. USDT->OUSD with OUSD contract
+    */
+    helperSwapETHWithOUSD: async function (destUser, ethAmountToSwap) {
+
+        ////////////// Loading some contracts //////////////
+
+        // loading WETH9 contract
+        const weth9 = new ethers.Contract(WETH9_ADDRESS, WETH9_ABI, destUser)
+        // loading USDT contract
+        const usdtToken = new ethers.Contract(USDT_ADDRESS, USDT_ABI, destUser)
+        // loading Tripool2 contract
+        const triPool = new ethers.Contract(CURVE_TRIPOOL2_ADDRESS, CURVE_TRIPOOL2_ABI, destUser)
+        // loading OUSD token contract
+        const ousdToken = new ethers.Contract(OUSD_ADDRESS, OUSD_ABI, destUser)
+        // loading OUSD Swapper contract
+        const ousdSwapper = new ethers.Contract(OUSD_SWAP_ADDRESS, OUSD_SWAP_ABI, destUser)
+
+        // Verify we got the correct TriPool connected (verifying USDT and WETH addresses)
+        ret = await triPool.coins(USDT_POOL_INDEX_NUMBER)
+        expect(ret).to.equal(USDT_ADDRESS)
+        ret = await triPool.coins(WETH_POOL_INDEX_NUMBER)
+        expect(ret).to.equal(WETH9_ADDRESS)
+
+        ////////////// 1. ETH->WETH9 //////////////
+
+        // read current signer balance from WETH9 contract (so we can validate increase later)
+        weth9Balance = await weth9.balanceOf(destUser.address)
+
+        // ETH->WETH @ WETH9 (becuase looks like tripool only deals with WETH)    
+        await weth9.deposit({value: ethAmountToSwap})
+
+        // read balance again and make sure it increased
+        expect(await weth9.balanceOf(destUser.address)).to.gt(weth9Balance)  
+        weth9Balance = await weth9.balanceOf(destUser.address)
+
+        ////////////// 2. WETH->USDT //////////////
+
+        // approve tripool to spend WETH9 on behalf of destUser
+        await weth9.approve(CURVE_TRIPOOL2_ADDRESS, ethAmountToSwap)
+
+        // get user balance
+        usdtBalance = await usdtToken.balanceOf(destUser.address)
+
+        // Exchange WETH9->USDT
+        // See: https://curve.readthedocs.io/factory-pools.html?highlight=exchange#StableSwap.exchange
+        // exchange(i: int128, j: int128, dx: uint256, min_dy: uint256, _receiver: address = msg.sender) → uint256: nonpayable
+        // i: Index value of the token to send.
+        // j: Index value of the token to receive.
+        // dx: The amount of i being exchanged.
+        // min_dy: The minimum amount of j to receive. If the swap would result in less, the transaction will revert.
+        await triPool.exchange(WETH_POOL_INDEX_NUMBER, USDT_POOL_INDEX_NUMBER, ethAmountToSwap, 1)
+
+        // read balance again and make sure it increased
+        expect(await usdtToken.balanceOf(destUser.address)).to.gt(usdtBalance)
+        usdtBalance = await usdtToken.balanceOf(destUser.address)
+    
+        ////////////// 3. USDT->OUSD with OUSD contract //////////////
+
+        // approve OUSD Swapper to spend USDT on behalf of destUser
+        await usdtToken.approve(OUSD_SWAP_ADDRESS, usdtBalance)
+
+        // get user balance
+        ousdBalance = await ousdToken.balanceOf(destUser.address)
+        
+        // Exchange USDT->OUSD
+        // https://github.com/OriginProtocol/origin-dollar/blob/e9ef066ab5cd8842d044e7759b99d956a44acd47/contracts/contracts/flipper/Flipper.sol
+        await ousdSwapper.buyOusdWithUsdt(usdtBalance)
+
+        // read balance again and make sure it increased
+        expect(await ousdToken.balanceOf(destUser.address)).to.gt(ousdBalance)
+        ousdBalance = await ousdToken.balanceOf(destUser.address)   
+    },
+    CURVE_TRIPOOL2_ADDRESS,
+    USDT_ADDRESS,
+    WETH9_ADDRESS,
+    CURVE_FACTORY_ADDRESS,
+    CURVE_3POOL_ADDRESS,
+    OUSD_SWAP_ADDRESS,
+    OUSD_ADDRESS,
+    USDT_POOL_INDEX_NUMBER,
+    WETH_POOL_INDEX_NUMBER,
+    OUSD_ABI,
+    OUSD_SWAP_ABI,
+    CURVE_TRIPOOL2_ABI,
+    USDT_ABI,
+    WETH9_ABI,
+    CURVE_FACTORY_ABI,
+}
+
