+++ test/zIntegrationTests.js
@@ -1,4 +1,4 @@
-var helper = require('./MainnerHelper');
+var helper = require('./MainnetHelper');
 
 const {BigNumber, FixedFormat, FixedNumber, formatFixed, parseFixed} = require("@ethersproject/bignumber");
 const { expect } = require("chai");
